python - test the BPs separately from ex1-ex3
ex1-ex3 only need libCEED, so moved the BP1 tests out of ex_test.py into their
own bps_test.py, which skips itself when petsc4py is not available. The test
target runs both.
